Move source out of boundary module and set SourceWavespectra as an entry point
