+++ rompy/core/data.py
@@ -17,8 +17,7 @@ from rompy.core.filters import Filter
 from rompy.core.grid import BaseGrid, RegularGrid
 from rompy.core.time import TimeRange
 from rompy.core.types import DatasetCoords, RompyBaseModel, Slice
-# from rompy.settings import DATA_SOURCE_TYPES
-# from rompy.utils import process_setting
+
 
 logger = logging.getLogger(__name__)
 
@@ -101,14 +100,8 @@ class DataBlob(RompyBaseModel):
 
 GRID_TYPES = Union[BaseGrid, RegularGrid]
 
-# DATA_SOURCE_TYPES = process_setting(DATA_SOURCE_TYPES)
-
 # Plugin for the source types
 SOURCE_TYPES = tuple([eps.load() for eps in entry_points(group="rompy.source")])
-# We could move these out of the module and specify them also with the entry-point
-# DATA_SOURCE_TYPES = (SourceDataset, SourceFile, SourceIntake, SourceDatamesh)
-# Append any additional data source types from entry points
-# DATA_SOURCE_TYPES = DATA_SOURCE_TYPES + tuple(eps.load() for eps in data_source_eps)
 
 
 class DataGrid(DataBlob):
